Updated Screen method; added several lvl objects
Screen methods such as line rendering, filled rectangle, and sprite
tiling. Level objects are no further advanced than being drawn on
screen. The next focus is actually creating a level editor to package
levels into data files, and to make it loads easier to create levels.

Graphics are interchangeable, and there will be support to include ones
own graphics, but as of yet there is no real user interface save some
drawstring diagnostic information, and some keybindings. An enhanced
menu system, and the implementation of entities/physics will all be
coming later. Level editing first.

My rendering engine is not efficient- objects are rendering over each
other, so its not like a tile map that doesn't allow any terrain
features to overlap (nice thing about tilemaps is it actually lets me
use less memory, since I only need to populate an array with many of the
same objects).

In the future, as an optimization measure, I may create a system to
perhaps cache an image of terrain for rendering, since it isn't really
necessary to put in a rendering request for static objects. I will still
have to store where objects are for actual gameplay, but I only need a
cached image for rendering. A PNG would do for all of the static stuff.
I wonder if I can fit a PNG into my format... hmm..

And for a closing message, parallax scrolling is like, tough.